feat: merge constants
