--- crates/kiwi-talk-app/src/app/constants.rs
@@ -1,11 +1,0 @@
-pub const TALK_VERSION: &str = "3.4.7";
-pub const TALK_OS: &str = "win32";
-pub const TALK_MCCMNC: &str = "999";
-pub const TALK_MODEL: &str = "";
-pub const TALK_NET_TYPE: i16 = 0;
-pub const TALK_DEVICE_TYPE: i8 = 2;
-pub const TALK_USE_SUB: bool = true;
-
-pub const BOOKING_SERVER: (&str, u16) = ("booking-loco.kakao.com", 443);
-
-pub const CHECKIN_SERVER: (&str, u16) = ("ticket-loco.kakao.com", 443);
